requirements.txt
python-docx
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
requests